Address review: relocate is_anthropic_model and tighten structured run
- Move is_anthropic_model from constants/llm.py to utils/llm.py per
  CLAUDE.md (utility functions belong in opencontractserver/utils/).
  Update imports in pydantic_ai_agents.py, data_extract_tasks.py, and
  the failure-classification test module.
- Trim verbose docstrings/comments in constants/llm.py while keeping the
  TOOL_LOOP_THRESHOLD ↔ STRUCTURED_OUTPUT_RETRIES coupling note explicit.
- Pass model_settings=None (not {}) to PydanticAIAgent when no
  temperature/max_tokens pin is set, preserving the pre-issue-#1381
  contract for non-Anthropic structured runs without caller pins.
  Add a regression test in test_pydantic_ai_agents.py.
- Convert the if/if/if return chain in _failure_message_for_classification
  to if/elif for clarity.
- Add an inline TODO at the doc_extract_query_task call site so a future
  refactor that accepts a caller-supplied model passes the same value to
  _resolve_extract_temperature, keeping model and temperature in lock-step.
